Make integer explicitly integer*4 in mex files
Start to add tests of spe, cut and slice objects

refs pace-neutrons/Herbert#13 #5

git-svn-id: https://svn.isis.rl.ac.uk/Herbert/trunk@76 6b2803e2-fb1c-43d7-acfb-f5118c1e7254

+++ external_code/Fortran/source_mex/bin_boundaries_from_descriptor.for
@@ -21,11 +21,11 @@
 
       implicit none
       mwPointer plhs(*), prhs(*)
-      integer nrhs, nlhs
+      integer*4 nrhs, nlhs
 
 ! mx routine declarations
       mwPointer mxCreateDoubleMatrix, mxGetPr
-      integer mxIsNumeric
+      integer*4 mxIsNumeric
       mwSize mxGetM, mxGetN
 
 ! Internal declations
@@ -34,7 +34,7 @@
       mwSize nx, nb, mx
       
 ! Arguments for computational routine, or purely internal
-      integer ierr, nx_pass, nb_pass, mx_pass
+      integer*4 ierr, nx_pass, nb_pass, mx_pass
       character*10 ch_num
       character*80 mess
