+++ external_code/Fortran/source_mex/rebin_2d_y_hist.for
@@ -24,11 +24,11 @@
 
       implicit none
       mwPointer plhs(*), prhs(*)
-      integer nrhs, nlhs
+      integer*4 nrhs, nlhs
 
 ! mx routine declarations
       mwPointer mxCreateDoubleMatrix, mxGetPr
-      integer mxIsNumeric
+      integer*4 mxIsNumeric
       mwSize mxGetM, mxGetN
 
 ! Internal declations
@@ -37,7 +37,7 @@
       mwSize nx, ny, my
       
 ! Arguments for computational routine, or purely internal
-      integer ierr, nx_pass, ny_pass, my_pass
+      integer*4 ierr, nx_pass, ny_pass, my_pass
       character*10 ch_num
       character*80 mess
